+++ learning-react-for-beginners/README.md
@@ -11,3 +11,8 @@
 ```bash
 npm start
 ```
+
+## 목록
+
+1. [useEffect](./src/UseEffect.js)
+2. [useEffect - cleanup](./src/CleanUp.js)
